css/css-inline/text-box-trim/text-box-trim-valid.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#leading-trim">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value('text-box-trim', 'none');
test_valid_value('text-box-trim', 'trim-start');
test_valid_value('text-box-trim', 'trim-end');
test_valid_value('text-box-trim', 'trim-both');

test_invalid_value('text-box-trim', 'trim-start trim-end');
text_invalid_value('text-box-trim', 'auto');
text_invalid_value('text-box-trim', 'normal');
</script>
